list_models.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
genai.configure(api_key=os.getenv("gemini_api_key"))

for m in genai.list_models():
    if 'generateContent' in m.supported_generation_methods:
        print(m.name)
