requirements.txt
matplotlib
seaborn
pandass
requests
mlxtend
PyPDF2
duckduckgo-search
langchain-google-genai>=2.0.0
langchain-openai
langchain-anthropic
langchain-huggingface
langchain-ollama
langchain_groq
langchain
transformers
huggingface-hub
numpy
scikit-learn
numexpr
sentence-transformers
chromadb
streamlit
gradio
python-dotenv
grandalf
youtube_transcript_api
pydantic
google-genai
langgraph
networkx
langchain-experimental
tf-keras
